==> notebooks/2-CombineFightGameData.ipynb <==
import requests
import json
import pandas as pd
import numpy as np

years = [ 20152016, 20162017, 20172018, 20182019, 20192020, 20202021]

frames = []
for year in years:
    df = pd.read_csv(f'AllFightGameData-{year}.csv')
    frames.append(df)
result = pd.concat(frames)
result.to_csv('AllFightGameDataCombined.csv', index=False)
