==> notebooks/4-FightAnalytics.ipynb <==
import numpy as np
import pandas as pd
import math

'''
Import the data
'''
df = pd.read_csv('FightAnalytics.csv')

'''
Find rows where post-total hits > pre-total hits
'''
more_total_hits = df.index[ (df['postHitsTotal'] > df['preHitsTotal']) & (df['preHitsTotal'] > 0) & (df['postHitsTotal'] > 0)].tolist()
more_total_penalties = df.index[ (df['postPenaltiesTotal'] > df['prePenaltiesTotal'])  & (df['postPenaltiesTotal'] > 0) & (df['prePenaltiesTotal'] > 0)].tolist()
more_total_shots = df.index[ (df['postShotsTotal'] > df['preShotsTotal']) & (df['postShotsTotal'] > 0) & (df['preShotsTotal'] > 0)].tolist()
hit_percent = (len(more_total_hits) / df.shape[0]) * 100
pen_percent = (len(more_total_penalties) / df.shape[0]) * 100
shot_percent = (len(more_total_shots) / df.shape[0]) * 100
print( "More hits " + "%.2f" % round(hit_percent, 2) + "% of the time")
print( "More penalties " + "%.2f" % round(pen_percent, 2) + "% of the time") 
print( "More shots " + "%.2f" % round(shot_percent, 2) + "% of the time") 

'''
Do the same for the home team
'''
more_home_hits = df.index[ (df['postHitsHome'] > df['preHitsHome']) & (df['preHitsHome'] > 0) & (df['postHitsHome'] > 0)].tolist()
more_home_penalties = df.index[ (df['postPenaltiesHome'] > df['prePenaltiesHome'])  & (df['postPenaltiesHome'] > 0) & (df['prePenaltiesHome'] > 0)].tolist()
more_home_shots = df.index[ (df['postShotsHome'] > df['preShotsHome']) & (df['postShotsHome'] > 0) & (df['preShotsHome'] > 0)].tolist()
home_hit_percent = (len(more_home_hits) / df.shape[0]) * 100
home_pen_percent = (len(more_home_penalties) / df.shape[0]) * 100
home_shot_percent = (len(more_home_shots) / df.shape[0]) * 100
print( "More home hits " + "%.2f" % round(home_hit_percent, 2) + "% of the time")
print( "More home penalties " + "%.2f" % round(home_pen_percent, 2) + "% of the time") 
print( "More home shots " + "%.2f" % round(home_shot_percent, 2) + "% of the time") 

'''
Do the same for the away team
'''
more_away_hits = df.index[ (df['postHitsAway'] > df['preHitsAway']) & (df['preHitsAway'] > 0) & (df['postHitsAway'] > 0)].tolist()
more_away_penalties = df.index[ (df['postPenaltiesAway'] > df['prePenaltiesAway'])  & (df['postPenaltiesAway'] > 0) & (df['prePenaltiesAway'] > 0)].tolist()
more_away_shots = df.index[ (df['postShotsAway'] > df['preShotsAway']) & (df['postShotsAway'] > 0) & (df['preShotsAway'] > 0)].tolist()
away_hit_percent = (len(more_away_hits) / df.shape[0]) * 100
away_pen_percent = (len(more_away_penalties) / df.shape[0]) * 100
away_shot_percent = (len(more_away_shots) / df.shape[0]) * 100
print( "More away hits " + "%.2f" % round(away_hit_percent, 2) + "% of the time")
print( "More away penalties " + "%.2f" % round(away_pen_percent, 2) + "% of the time") 
print( "More away shots " + "%.2f" % round(away_shot_percent, 2) + "% of the time")
